+++ openpyxl/cell/tests/test_read_only.py
@@ -4,7 +4,7 @@ import datetime
 import pytest
 
 from openpyxl.cell.read_only import ReadOnlyCell
-from openpyxl.collections import IndexedList
+from openpyxl.utils.indexed_list import IndexedList
 
 
 @pytest.fixture(scope='module')
@@ -12,6 +12,10 @@ def dummy_sheet():
     class DummyWorkbook(object):
         shared_styles = IndexedList()
         shared_styles.add(None) # Workbooks always have a default style
+        _cell_styles = IndexedList()
+        _cell_styles.add(None)
+        _number_formats = IndexedList()
+
 
     class DummySheet(object):
         base_date = 2415018.5
@@ -82,14 +86,13 @@ def test_numeric(dummy_sheet, value, expected):
 
 @pytest.fixture(scope="class")
 def DummyCell(dummy_sheet):
-    class DummyNumberFormat:
-        format_code = 'd-mmm-yy'
 
     class DummyStyle(object):
-        number_format = 'd-mmm-yy'
+        number_format = 164
 
-    dummy_sheet.parent.shared_styles.add(DummyStyle())
-    cell = ReadOnlyCell(dummy_sheet, None, None, "23596", 'n', '1')
+    dummy_sheet.parent._number_formats.add('d-mmm-yy')
+    dummy_sheet.parent._cell_styles.add(DummyStyle())
+    cell = ReadOnlyCell(dummy_sheet, None, None, "23596", 'n', DummyStyle)
     return cell
 
 
@@ -109,7 +112,8 @@ class TestDateTime:
 
 
 def test_read_only():
-    cell = ReadOnlyCell(None, None, 1, None)
+    cell = ReadOnlyCell(sheet=None, row=None, column=None, value=1)
+    assert cell.value == 1
     with pytest.raises(AttributeError):
         cell.value = 10
     with pytest.raises(AttributeError):