+++ openpyxl/workbook/names/named_range.py
@@ -8,9 +8,9 @@ import re
 import warnings
 
 # package imports
-from openpyxl.cell import absolute_coordinate
+from openpyxl.utils import absolute_coordinate
 from openpyxl.compat import unicode
-from openpyxl.exceptions import NamedRangeException
+from openpyxl.utils.exceptions import NamedRangeException
 from openpyxl.xml.functions import fromstring, safe_iterator
 from openpyxl.xml.constants import SHEET_MAIN_NS