+++ openpyxl/writer/worksheet.py
@@ -11,10 +11,10 @@ from io import BytesIO
 from openpyxl.compat import safe_string, itervalues
 
 # package imports
+from openpyxl.utils import COORD_RE
 from openpyxl.cell import (
     coordinate_from_string,
     column_index_from_string,
-    COORD_RE
 )
 from openpyxl.xml.functions import (
     XMLGenerator,
@@ -33,6 +33,7 @@ from openpyxl.xml.constants import (
 from openpyxl.compat.itertools import iteritems, iterkeys
 from openpyxl.formatting import ConditionalFormatting
 from openpyxl.worksheet.datavalidation import writer
+from openpyxl.worksheet.properties import WorksheetProperties, write_sheetPr
 
 
 def row_sort(cell):
@@ -40,14 +41,9 @@ def row_sort(cell):
     return column_index_from_string(cell.column)
 
 
-def write_properties(worksheet, vba_attrs):
-    pr = Element('sheetPr', vba_attrs)
-    summary = Element('outlinePr',
-                      summaryBelow='%d' % worksheet.show_summary_below,
-                      summaryRight= '%d' % worksheet.show_summary_right)
-    pr.append(summary)
-    if worksheet.page_setup.fitToPage:
-        pr.append(Element('pageSetUpPr', fitToPage='1'))
+def write_properties(worksheet):
+    wsp = worksheet.sheet_properties
+    pr = write_sheetPr(wsp)
     return pr
 
 
@@ -57,13 +53,16 @@ def write_sheetviews(worksheet):
     if not worksheet.show_gridlines:
         sheetviewAttrs['showGridLines'] = '0'
     view = SubElement(views, 'sheetView', sheetviewAttrs)
-    selectionAttrs = {}
+    selectionAttrs = {
+        'activeCell': worksheet.active_cell,
+        'sqref': worksheet.selected_cell
+    }
     topLeftCell = worksheet.freeze_panes
     if topLeftCell:
         colName, row = coordinate_from_string(topLeftCell)
         column = column_index_from_string(colName)
         pane = 'topRight'
-        paneAttrs = {}
+        paneAttrs = {'topLeftCell':topLeftCell, 'state':'frozen'}
         if column > 1:
             paneAttrs['xSplit'] = str(column - 1)
         if row > 1:
@@ -71,17 +70,12 @@ def write_sheetviews(worksheet):
             pane = 'bottomLeft'
             if column > 1:
                 pane = 'bottomRight'
-        paneAttrs.update(dict(topLeftCell=topLeftCell,
-                              activePane=pane,
-                              state='frozen'))
-        view.append(Element('pane', paneAttrs))
         selectionAttrs['pane'] = pane
+        paneAttrs['activePane'] = pane
+        view.append(Element('pane', paneAttrs))
         if row > 1 and column > 1:
-            SubElement(view, 'selection', {'pane': 'topRight'})
-            SubElement(view, 'selection', {'pane': 'bottomLeft'})
-
-    selectionAttrs.update({'activeCell': worksheet.active_cell,
-                           'sqref': worksheet.selected_cell})
+            SubElement(view, 'selection', pane='topRight')
+            SubElement(view, 'selection', pane='bottomLeft')
 
     SubElement(view, 'selection', selectionAttrs)
     return views
@@ -144,7 +138,7 @@ def write_autofilter(worksheet):
             for val in filter_column.vals:
                 flt.append(Element('filter', val=val))
         if auto_filter.sort_conditions:
-            srt = SubElement(el,  'sortState', ref=auto_filter.ref)
+            srt = SubElement(el, 'sortState', ref=auto_filter.ref)
             for sort_condition in auto_filter.sort_conditions:
                 sort_attr = {'ref': sort_condition.ref}
                 if sort_condition.descending:
@@ -245,7 +239,7 @@ def write_hyperlinks(worksheet):
 def write_pagebreaks(worksheet):
     breaks = worksheet.page_breaks
     if breaks:
-        tag = Element( 'rowBreaks', {'count': str(len(breaks)),
+        tag = Element('rowBreaks', {'count': str(len(breaks)),
                                      'manualBreakCount': str(len(breaks))})
         for b in breaks:
             tag.append(Element('brk', id=str(b), man=true, max='16383',
@@ -262,7 +256,7 @@ def write_worksheet(worksheet, shared_strings):
               {'xmlns': SHEET_MAIN_NS,
                'xmlns:r': REL_NS})
 
-    props = write_properties(worksheet, worksheet.vba_code)
+    props = write_properties(worksheet)
     xml_file.write(tostring(props))
 
     dim = Element('dimension', {'ref': '%s' % worksheet.calculate_dimension()})
@@ -378,6 +372,8 @@ def write_worksheet_data(doc, worksheet):
         start_tag(doc, 'row', attrs)
         row_cells = cells_by_row[row_idx]
         for cell in sorted(row_cells, key=row_sort):
+            if cell.value is None and not cell.has_style:
+                continue
             write_cell(doc, worksheet, cell)
 
         end_tag(doc, 'row')
@@ -392,7 +388,7 @@ def write_cell(doc, worksheet, cell):
     coordinate = cell.coordinate
     attributes = {'r': coordinate}
     if cell.has_style:
-        attributes['s'] = '%d' % cell._style
+        attributes['s'] = '%d' % cell.style_id
 
     if cell.data_type != cell.TYPE_FORMULA:
         attributes['t'] = cell.data_type